add placeholder event pages
