+++ src/locale/misc/breadcrumbs/en.yml
@@ -5,3 +5,4 @@ projects: Projects
 areas: Areas
 campaigns: Campaigns
 insights: Insights
+events: Events
